requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
brotli==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
Faker==37.12.0
fonttools==4.60.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
joblib==1.5.2
multidict==6.7.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
postgrest==2.24.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
 xhtml2pdf
realtime==2.24.0
reportlab==4.4.4
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.24.0
StrEnum==0.4.15
stripe==13.2.0
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
zopfli==0.4.0
